requirements.txt
numpy
nltk
h5py
joblib
fairseq==0.9.0
pytorch_pretrained_bert
torch_scatter
networkx
jsonpickle~=2.0.0
lxml
sklearn
tabulate